Improve gravedigger code
